+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * added `AsyncDecoratorTrait` to ease processing responses without breaking async
  * added support for pausing responses with a new `pause_handler` callable exposed as an info item
  * added `StreamableInterface` to ease turning responses into PHP streams
+ * added `MockResponse::getRequestMethod()` and `getRequestUrl()` to allow inspecting which request has been sent
 
 5.1.0
 -----